CodeGraphAI/requirements.txt
# CodeGraphAI - Dependências Principais
# Python 3.8+

# Bancos de Dados (Opcional - instale apenas os necessários)
# Oracle
oracledb>=1.4.0

# PostgreSQL
psycopg2-binary>=2.9.0

# SQL Server (escolha um dos drivers)
pyodbc>=5.0.0
# ou
# pymssql>=2.2.0

# MySQL (escolha um dos drivers)
mysql-connector-python>=8.0.0
# ou
# pymysql>=1.0.0

# SQLAlchemy (opcional, para abstração futura)
# sqlalchemy>=2.0.0

# LangChain - Framework para LLM
langchain>=0.1.0
langchain-community>=0.0.13

# Transformers e PyTorch - Modelos de IA
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # Para quantização 8-bit (opcional)

# Análise de Grafos
networkx>=3.0

# Visualização
matplotlib>=3.7.0

# CLI
click>=8.0.0

# Progress tracking (será usado na Fase 5.4)
tqdm>=4.65.0

# Configuração via .env (será usado na Fase 2.2)
python-dotenv>=1.0.0

